+++ core/src/com/bears/drift/Tile.java
@@ -9,6 +9,6 @@ public abstract class Tile {
 
     abstract void fun();
 
-    abstract int distanceToCurve1(int x, int y);
-    abstract int distanceToCurve2(int x, int y);
+    abstract float distanceToCurve1(float x, float y);
+    abstract float distanceToCurve2(float x, float y);
 }
